etl/requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
certifi==2024.2.2
distlib==0.3.8
elastic-transport==8.12.0
elasticsearch==8.12.1
Faker==22.6.0
filelock==3.13.1
flake8==6.1.0
isort==5.13.2
mccabe==0.7.0
platformdirs==4.2.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pydantic==2.6.3
pydantic-settings==2.2.1
pydantic_core==2.16.3
pyflakes==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.1